+++ src/main/java/com/zerobase/gamecollectors/controller/SignOutController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(value = "/sign-out")
+@RequestMapping("/sign-out")
 @RequiredArgsConstructor
 public class SignOutController {
