+++ src/main/java/com/zerobase/gamecollectors/service/ManagerSignInOutService.java
@@ -8,8 +8,8 @@ import com.zerobase.gamecollectors.domain.entity.Manager;
 import com.zerobase.gamecollectors.domain.repository.ManagerRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.SignInServiceDto;
 import com.zerobase.gamecollectors.model.TokenDto;
+import com.zerobase.gamecollectors.model.sign.SignInServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import java.util.Date;
 import javax.transaction.Transactional;
