+++ src/main/java/com/zerobase/gamecollectors/service/ManagerSignUpService.java
@@ -5,8 +5,8 @@ import com.zerobase.gamecollectors.domain.entity.Manager;
 import com.zerobase.gamecollectors.domain.repository.ManagerRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.ManagerSignUpServiceDto;
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.sign.ManagerSignUpServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import com.zerobase.gamecollectors.util.RandomCodeGenerator;
 import javax.transaction.Transactional;
