+++ src/main/java/com/zerobase/gamecollectors/service/SendEmailService.java
@@ -1,7 +1,7 @@
 package com.zerobase.gamecollectors.service;
 
 import com.zerobase.gamecollectors.client.mailgun.MailgunClient;
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
