+++ src/main/java/com/zerobase/gamecollectors/service/UserSignUpService.java
@@ -9,8 +9,8 @@ import com.zerobase.gamecollectors.domain.repository.PointRepository;
 import com.zerobase.gamecollectors.domain.repository.UserRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
-import com.zerobase.gamecollectors.model.UserSignUpServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.sign.UserSignUpServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import com.zerobase.gamecollectors.util.RandomCodeGenerator;
 import javax.transaction.Transactional;
