+++ src/test/java/com/zerobase/gamecollectors/service/ManagerSignUpServiceTest.java
@@ -17,8 +17,8 @@ import com.zerobase.gamecollectors.domain.entity.Manager;
 import com.zerobase.gamecollectors.domain.repository.ManagerRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.ManagerSignUpServiceDto;
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.sign.ManagerSignUpServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
