+++ src/test/java/com/zerobase/gamecollectors/service/SendEmailServiceTest.java
@@ -2,7 +2,7 @@ package com.zerobase.gamecollectors.service;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
 import java.util.Objects;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
