+++ src/test/java/com/zerobase/gamecollectors/service/UserSignInOutServiceTest.java
@@ -18,8 +18,8 @@ import com.zerobase.gamecollectors.domain.entity.User;
 import com.zerobase.gamecollectors.domain.repository.UserRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.SignInServiceDto;
 import com.zerobase.gamecollectors.model.TokenDto;
+import com.zerobase.gamecollectors.model.sign.SignInServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import java.util.Date;
 import java.util.Optional;
