+++ src/test/java/com/zerobase/gamecollectors/service/UserSignUpServiceTest.java
@@ -21,8 +21,8 @@ import com.zerobase.gamecollectors.domain.repository.PointRepository;
 import com.zerobase.gamecollectors.domain.repository.UserRepository;
 import com.zerobase.gamecollectors.exception.CustomException;
 import com.zerobase.gamecollectors.exception.ErrorCode;
-import com.zerobase.gamecollectors.model.SendEmailServiceDto;
-import com.zerobase.gamecollectors.model.UserSignUpServiceDto;
+import com.zerobase.gamecollectors.model.email.SendEmailServiceDto;
+import com.zerobase.gamecollectors.model.sign.UserSignUpServiceDto;
 import com.zerobase.gamecollectors.redis.RedisUtil;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
